+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Features
 
+  - Styling: checkboxes uses `form-checkbox` class.
+  - Styling: all layouts now adds `form-group` class.
   - Styling: required fields get a `required` class.
 
 # 1.2.3